+++ CHANGELOG.md
@@ -6,7 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## CultureInfo.js [Unreleased]
 
-[Show differences](https://github.com/manuth/CultureInfo.js/compare/v1.0.11...dev)
+[Show differences](https://github.com/manuth/CultureInfo.js/compare/v2.0.0...dev)
+
+## CultureInfo.js v2.0.0
+### Breaking
+  - The `CultureInfo.IsNeutralCulture` now considers all cultures which don't have a region specified, as neutral (`zh-Hans` or `zh-Hant`, but also `de` or `en`)
+
+### Fixed
+  - Vulnerabilities in dependencies
+
+### Added
+  - Support for the Test Explorer UI
+  - Support for the `ts-nameof` plugin
+  - New unit-tests
+
+### Updated
+  - All dependencies
+
+[Show differences](https://github.com/manuth/CultureInfo.js/compare/v1.0.11...v2.0.0)
 
 ## CultureInfo.js v1.0.11
 ### Fixed